Добавил в попап ошибки код ошибки, функции сортировки перенес в util
