+++ index.html
@@ -256,7 +256,8 @@
   <!-- Сообщение об ошибке при загрузке данных -->
   <!-- error-server--open - показать ошибку -->
   <div class="error-server">
-    <p class="error-server__message">Ошибка при загрузке данных с сервера</p>
+    <p class="error-server__title">Ошибка при загрузке данных с сервера</p>
+    <p class="error-server__message"></p>
     <button class="error-server__button" type="button" aria-label="Закрыть уведомление об ошибке"></button>
   </div>
